requirements.txt
Django
psycopg2
social-auth-app-django
django-taggit
Pillow
django-imagekit
django-cleanup

